services/loan-service/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
django-simple-history==3.0.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
django-cors-headers==3.7.0
python-decouple==3.6
gunicorn==20.1.0
